Android Circles App
This is an android app for handling user touch events using 2D graphics
and accelerometer.

User’s touch event is used to generate circles on the screen. The
circle can be moved in the desired direction with desired velocity.

Supports collision detection and gravity. Used accelerometer to modify
the direction and velocity of circles.